+++ lib/elm_common/example/.gitignore
@@ -1,0 +1,5 @@
+elm-stuff/
+Hawk.elm
+Redirect.elm
+TaskclusterLogin.elm
+User.elm
