+++ lib/elm_common/example/index.js
@@ -1,0 +1,6 @@
+var app = Elm.Main.fullscreen();
+
+console.log('Run App', app);
+window.localstorage(app, 'bugzilla');
+window.localstorage(app, 'taskclusterlogin');
+window.redirect(app);
